ルーンを小数点で登録できるよう修正

+++ rune/app/Http/Controllers/HomeController.php
@@ -38,7 +38,7 @@ class HomeController extends Controller
                 $sumRune += $helpHistory->rune;
             }
 
-            $yen = (int)$sumRune * (int)$runeRate->rate;
+            $yen = (float)$sumRune * (float)$runeRate->rate;
 
             return view('home', compact('runeRate','sumRune','yen'));
         }
